Still buggy and incomplete, but added basic entity storage, which lets it cache players, so the position isn't client side.

+++ src/main/java/omnivoxel/server/entity/mob/MobEntity.java
@@ -26,7 +26,7 @@ public abstract class MobEntity extends GravityEntity {
         return yaw;
     }
 
-    public void setYaw(float yaw) {
+    public void setYaw(double yaw) {
         this.yaw = yaw;
     }
 
@@ -34,7 +34,7 @@ public abstract class MobEntity extends GravityEntity {
         return pitch;
     }
 
-    public void setPitch(float pitch) {
+    public void setPitch(double pitch) {
         this.pitch = pitch;
     }
 }
